backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# WebSocket support
python-socketio==5.11.0
python-engineio==4.9.0

# Data validation and settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Existing dependencies from original project
streamlit
gradio_client
requests
mido

# Optional: Google Magenta and dependencies
# Note: These may need to be installed separately via conda
# magenta
# tensorflow-cpu
# pretty_midi
# librosa
# python-rtmidi
